+++ src/shared/GameInfo.h
@@ -35,6 +35,7 @@ struct GameInfo {
         SDL2_SURFACE = 0x04,
         SDL2_RENDERER = 0x08,
         OPENGL = 0x10,
+        VDPAU = 0x20,
         OPENAL = 0x100,
         PULSEAUDIO = 0x200,
         XEVENTS = 0x400,