+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Non-blocking ALSA mode (#361)
 * Add a non-interactive mode
 * Support window title change from XChangeProperty()
+* Add VDPAU rendering
 
 ### Changed
 ### Fixed